Section 4 part 1 edits
bold items, I did not edit. 
1366:  includes -> include
1482:  lower case primary and secondary and identifiers in this section (lines 1485, 1489, 1491
1488: paragraph space missing (I did a line space, but not sure if that took care of it). 
1513: table is in middle of 15F text.  It should be right after 2F
1517: need spacing between listed items for employee color scheme (bothh top and bottom)
1556: Incomplete sentence.  placement of what?   add:  of the ICC(s) and related components"
1563: do not allow -> MAY not allow
1569: may -> MAY
1571: may -> MAY
1574:  do not allow -> MAY not allow
1598: do not allow -> MAY not allow

1623: should-> SHOULD
1629: should -> SHOULD
1635:  may-> MAY
1656: may -> MAY
1690: since this subsection talks about authentication and lists the credentials, a footnote is appropriate.  add a footnote: The CHUID as an authentication mechanims (Section 6.2.5) has been
removed from this version of the Standard. The CHUID data element itself, however, has not been removed and continues to be mandatory as it supports other PIV authentication mechanisms.
1724:  This is the biggest section about the CHUID - yet, it does not state that the authentication mechanism is deprecated.  Suggest to add the following text.

The CHUID authentication mechanims (Section 6.2.5) has been removed from this version of the Standard. The CHUID data element, however, has not been removed and continues to be mandatory as it supports other PIV authentication mechanisms. For example, the BIO, BIO-A, and SYM-CAK authentication mechanisms use the CHUID data  element as a source for the card’s expiration date. The CHUID data element also provides the content signing certificate for some authentication mechanisms and unique identifiers for PACS ACLs. 
1800:  This para is missplaced. I moved it back (underneath para beginning with 1784