Update requirements.md
initiation of investigation ,as described in the prior footnote, is only partial requirements, per OPM. The request for investigation has to be submitted to the ISP AND ISP has  to have scheduled the investigation.  Hence, it is better to point to CSP where initiation of background investigation is clearly defined.